move all views into views/
this is where they belong :)
