--- scripts/test/tester.php
@@ -1,5 +1,0 @@
-<?php
-
-include('../dao.php');
-
-addDepartment("nathan");
